Hacker News improvement feedback
NaN safe cached inverse ray direction; Reduce passing std::shared_ptr by value; Use float instead of double where extra precision is not needed; Tests to cover this;

+++ include/luz/Sampler.hpp
@@ -8,8 +8,8 @@ namespace Sampler
 {
 	struct	Sample2D
 	{
-		double	x = 0.0;
-		double	y = 0.0;
+		float	x = 0.0f;
+		float	y = 0.0f;
 	};
 
 	enum	Dimension : std::uint32_t
